Main screen and view model upgated

+++ app/src/main/java/com/tes/apps/development/aptoideappdownloader/data/remote/dto/AppInfoDto.kt
@@ -3,19 +3,19 @@ package com.tes.apps.development.aptoideappdownloader.data.remote.dto
 
 import com.google.gson.annotations.SerializedName
 
-data class AppInfoDto(
+data class AppInfoDto (
     @SerializedName("added")
     val added: String = "",
     @SerializedName("apk_tags")
     val apkTags: List<String> = listOf(),
     @SerializedName("downloads")
-    val downloads: Int = 0,
+    val downloads: Long = 0,
     @SerializedName("graphic")
     val graphic: String = "",
     @SerializedName("icon")
     val icon: String = "",
     @SerializedName("id")
-    val id: Int = 0,
+    val id: Long = 0,
     @SerializedName("md5sum")
     val md5sum: String = "",
     @SerializedName("modified")
@@ -25,13 +25,13 @@ data class AppInfoDto(
     @SerializedName("package")
     val packageX: String = "",
     @SerializedName("pdownloads")
-    val pdownloads: Int = 0,
+    val pdownloads: Long = 0,
     @SerializedName("rating")
     val rating: Double = 0.0,
     @SerializedName("size")
-    val size: Int = 0,
+    val size: Long = 0,
     @SerializedName("store_id")
-    val storeId: Int = 0,
+    val storeId: Long = 0,
     @SerializedName("store_name")
     val storeName: String = "",
     @SerializedName("updated")
@@ -39,7 +39,7 @@ data class AppInfoDto(
     @SerializedName("uptype")
     val uptype: String = "",
     @SerializedName("vercode")
-    val vercode: Int = 0,
+    val vercode: Long = 0,
     @SerializedName("vername")
     val vername: String = ""
 )
